+++ taxsystem/task_helpers/payment_helpers.py
@@ -4,7 +4,8 @@ from django.db import transaction
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 
-from taxsystem.hooks import get_extension_logger
+from allianceauth.services.hooks import get_extension_logger
+
 from taxsystem.models.filters import SmartGroup
 from taxsystem.models.logs import PaymentHistory
 from taxsystem.models.tax import (